UI: Rename ComponentList to ComponentTree.
