+++ packages/ui-storybook/stories/index.tsx
@@ -1,5 +1,5 @@
 // tslint:disable-next-line:no-submodule-imports
 import "@prodo-ai/snoopy-ui/src/App/index.css";
 
-import "./ComponentList";
+import "./ComponentTree";
 import "./Sidebar";
