+++ packages/ui/examples/ComponentTree.example.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
-import ComponentList from "../src/components/ComponentList";
+import ComponentTree from "../src/components/ComponentTree";
 import {testComponents} from "../test/fixtures";
 
-export default ComponentList;
+export default ComponentTree;
 
 export const Empty = () => (
-  <ComponentList
+  <ComponentTree
     components={[]}
     select={() => {
       alert("select");
@@ -15,7 +15,7 @@ export const Empty = () => (
 );
 
 export const singleItem = () => (
-  <ComponentList
+  <ComponentTree
     components={[testComponents[0]]}
     select={() => {
       alert("select");
@@ -26,7 +26,7 @@ export const singleItem = () => (
 singleItem.title = "Single Item";
 
 export const singleItemWithSelection = () => (
-  <ComponentList
+  <ComponentTree
     components={[testComponents[0]]}
     selected={[testComponents[0].path]}
     select={() => {
@@ -37,7 +37,7 @@ export const singleItemWithSelection = () => (
 singleItemWithSelection.title = "Single item with selection";
 
 export const multipleItems = () => (
-  <ComponentList
+  <ComponentTree
     components={testComponents}
     select={() => {
       alert("select");
@@ -48,7 +48,7 @@ export const multipleItems = () => (
 multipleItems.title = "Multiple items";
 
 export const multipleItemsWithSelection = () => (
-  <ComponentList
+  <ComponentTree
     components={testComponents}
     selected={[testComponents[0].path]}
     select={() => {
